Average view added
